+++ R/dotplot_method.R
@@ -23,9 +23,9 @@ setMethod("dotplot", "features", function(.Object, col = NULL, n = 20L, ...){
 
 #' @rdname dotplot-method
 setMethod("dotplot", "featuresNgrams", function(.Object, col = NULL, n = 20L, ...){
-  if (is.null(col)) col <- ngrams_tex@method[1]
+  if (is.null(col)) col <- .Object@method[1]
   dt_subset <- .Object@stat[n:1L][,grep("^\\d+_.*$", colnames(.Object)), with = FALSE][,n := 1L:.N]
-  ngram_vector <- dt_subset[,paste(unlist(.SD[1,]), collapse = " "), by = .(n)][["V1"]]
+  ngram_vector <- dt_subset[,paste(unlist(.SD[1,]), collapse = " "), by = c("n")][["V1"]]
   dotchart(x = .Object@stat[[col]][n:1L], labels = ngram_vector, ...)
 })