+++ NEWS.md
@@ -5,6 +5,8 @@
 * dotplot-method added for featuresNgrams
 * sample corpus GermaParlMini added to the package (replacing suggested package polmineR.sampleCorpus)
 * configuration mechanism added to set path to data directory in registry file upon installation 
+* class hits now inherits from class 'textstat', exposing a set of generic functions (such as dim, nrow etc.); slot 'dt' changed to 'stat' for this purpose
+* count,partitionBundle and hits,partitionBundle: cqp parameter added
 
 # v0.7.6
 * functionality of matches-method (breakdown of frequencies of matches) integrated